requirements.txt
anthropic>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
youtube-transcript-api>=1.0.0
yt-dlp>=2024.1.0
redis>=5.0.0
rq>=1.16.0
pydantic>=2.0.0
tenacity>=8.2.0
loguru>=0.7.0
